Feature/ASE-109

+++ src/main/java/com/bloxico/ase/userservice/entity/user/Role.java
@@ -27,7 +27,7 @@ public class Role implements Serializable {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    private Short id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
