+++ src/main/java/com/bloxico/ase/userservice/web/model/evaluation/DeleteCountryEvaluationDetailsRequest.java
@@ -18,6 +18,6 @@ public class DeleteCountryEvaluationDetailsRequest {
     @NotNull
     @JsonProperty("id")
     @ApiModelProperty(required = true)
-    Integer id;
+    Long id;
 
 }
